refactor: share parsePath util
