+++ tests/unit/favorites.test.js
@@ -1,6 +1,6 @@
 /* global describe, test, expect, beforeEach, jest */
+import { parsePath } from "../../src/popup/path-utils.js";
 import {
-  parsePath,
   loadFavorites,
   importFavoritesFromText,
   exportFavorites,
